tests/wpt/tests/html/canvas/offscreen/webgpu-access/2d.webgpuaccess.getTextureFormat.rgba8.tentative.https.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.webgpuaccess.getTextureFormat.rgba8.tentative.https</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.webgpuaccess.getTextureFormat.rgba8.tentative.https</h1>
<p class="desc">getTextureFormat() returns RGBA8 or BGRA8 for a typical context</p>


<script>
var t = async_test("getTextureFormat() returns RGBA8 or BGRA8 for a typical context");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_regexp_match(ctx.getTextureFormat(), /^rgba8unorm|bgra8unorm$/);
  t.done();

});
</script>
